requirements.txt
tensorflow==2.3.0
matplotlib==3.2.2
tqdm==4.49.0
numpy==1.18.5
Pillow==7.2.0
scikit_learn==0.23.2
